+++ packages/form/form-item.vue
@@ -1,0 +1,47 @@
+<template>
+  <div class="form-group">
+    <label
+      v-if="label"
+      :for="$attrs.for"
+      class="block text-md font-medium text-gray-700"
+    >
+      {{ label }}
+    </label>
+    <slot :id="uId" name="default"></slot>
+    <p
+      v-if="desc"
+      id="email-description"
+      class="mt-2 text-xs text-gray-500"
+      v-html="desc"
+    ></p>
+  </div>
+</template>
+
+<script>
+import UniqueComponentId from '../utils/UniqueComponentId'
+
+export default {
+  name: 'IFormItem',
+  props: {
+    /** Label of Input */
+    label: { type: String, default: null },
+    /** Description text */
+    desc: { type: String, default: null },
+  },
+  data() {
+    return {
+
+    }
+  },
+  computed: {
+    uId() {
+      return UniqueComponentId()
+    },
+  },
+}
+</script>
+<style lang="stylus">
+.form-group {
+  @apply block mb-3;
+}
+</style>