+++ src/components/utils/UniqueComponentId.js
@@ -1,6 +1,0 @@
-var lastId = 0;
-
-export default function (prefix = 'i-uid-') {
-  lastId++;
-  return `${prefix}${lastId}`;
-}