+++ src/main.js
@@ -1,6 +1,0 @@
-import {createApp} from 'vue';
-import App from './App.vue';
-
-import iSPAElement from './index';
-
-createApp(App).use(iSPAElement).mount('#app');